refactor : 전체적인 코드 구조 개선
